// mlflow/server/js/vendor/design-system/src/theme/_generated/PrimitiveColors.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT EDIT.

import type { ValidPrimitiveColors } from './ValidPrimitiveColors';

export const primitiveColors: Record<ValidPrimitiveColors, string> = {
  blue100: '#F0F8FF',
  blue200: '#D7EDFE',
  blue300: '#BAE1FC',
  blue400: '#8ACAFF',
  blue500: '#4299E0',
  blue600: '#2272B4',
  blue700: '#0E538B',
  blue800: '#04355D',
  brown: '#A6630C',
  coral: '#C83243',
  green100: '#F3FCF6',
  green200: '#D4F7DF',
  green300: '#B1ECC5',
  green400: '#8DDDA8',
  green500: '#3CAA60',
  green600: '#277C43',
  green700: '#115026',
  green800: '#093919',
  grey050: '#F6F7F9',
  grey100: '#E8ECF0',
  grey200: '#D1D9E1',
  grey300: '#C0CDD8',
  grey350: '#92A4B3',
  grey400: '#8396A5',
  grey500: '#5F7281',
  grey600: '#445461',
  grey650: '#37444F',
  grey700: '#1F272D',
  grey800: '#11171C',
  indigo: '#434A93',
  lemon: '#FACB66',
  lime: '#308613',
  pink: '#B45091',
  purple: '#8A63BF',
  red100: '#FFF5F7',
  red200: '#FDE2E8',
  red300: '#FBD0D8',
  red400: '#F792A6',
  red500: '#E65B77',
  red600: '#C82D4C',
  red700: '#9E102C',
  red800: '#630316',
  teal: '#04867D',
  turquoise: '#137DAE',
  white: '#FFFFFF',
  yellow100: '#FFF9EB',
  yellow200: '#FCEACA',
  yellow300: '#F8D4A5',
  yellow400: '#F2BE88',
  yellow500: '#DE7921',
  yellow600: '#BE501E',
  yellow700: '#93320B',
  yellow800: '#5F1B02',
};